Recalculate fNcells of TH1
fSumw2.fN of TH1 depends on fNcells which is getting set in
constructors. There are only 2 ways to recalculate it for TH3 (+ some
shit from TH1, nvm):
1. TH3::Rebind3D()
2. TH3::SetBinsLength()

We're gonna use the second one with the negative argument to make it
recalculate fNcells based on the 3D axis number of bins.